requirements.txt
python-dotenv==1.0.1
python-telegram-bot==13.7
logzero==1.7.0
openai==1.23.6
tenacity~=8.2.2
black~=23.3.0
webdriver-manager~=4.0.0
selenium~=4.9.1